provenance/requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
gitdb==4.0.5
GitPython==3.1.3
h11==0.9.0
httptools==0.1.1
idna==2.9
pika==1.1.0
pony==0.7.14
python-dotenv==0.13.0
requests==2.23.0
smmap==3.0.4
starlette==0.13.4
unidiff==0.6.0
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
